+++ pkgs/jni/lib/src/core_bindings.dart
@@ -6806,15 +6806,14 @@ extension type JCharacter$JUnicodeScript._(jni$_.JObject _$this)
           jni$_.JniResult Function(jni$_.Pointer<jni$_.Void>,
               jni$_.JMethodIDPtr, jni$_.Pointer<jni$_.Void>)>();
 
-  /// from: `static public java.lang.Character$UnicodeScript valueOf(java.lang.String synthetic)`
+  /// from: `static public java.lang.Character$UnicodeScript valueOf(java.lang.String string)`
   /// The returned object must be released after use, by calling the [release] method.
   static JCharacter$JUnicodeScript? valueOf(
-    JString? synthetic,
+    JString? string,
   ) {
     final _$$classRef = _class.reference;
-    final _$synthetic = synthetic?.reference ?? jni$_.jNullReference;
-    return _valueOf(
-            _$$classRef.pointer, _id_valueOf.pointer, _$synthetic.pointer)
+    final _$string = string?.reference ?? jni$_.jNullReference;
+    return _valueOf(_$$classRef.pointer, _id_valueOf.pointer, _$string.pointer)
         .object<JCharacter$JUnicodeScript?>();
   }
 
@@ -10297,7 +10296,7 @@ extension type JFloat._(jni$_.JObject _$this)
   static const MIN_EXPONENT = -126;
 
   /// from: `static public final float MIN_NORMAL`
-  static const MIN_NORMAL = 1.17549435e-38;
+  static const MIN_NORMAL = 1.1754944e-38;
 
   /// from: `static public final float MIN_VALUE`
   static const MIN_VALUE = 1.4e-45;