+++ pkgs/jni/src/dartjni.h
@@ -15,10 +15,8 @@
 
 #if defined(__ANDROID__)
 #include <android/log.h>
-#define __ENVP_CAST (JNIEnv**)
 #define LOGF(...) __android_log_print(ANDROID_LOG_FATAL, "DartJNI", __VA_ARGS__)
 #else
-#define __ENVP_CAST (void**)
 #define LOGF(...)                                                              \
   do {                                                                         \
     fprintf(stderr, "DartJNI [FATAL] ");                                       \
@@ -165,7 +163,7 @@ static inline void attach_thread() {
           " Dart code too early, before 'main()' (such as during Dart plugin"
           " class registration)?");
     }
-    (*jni->jvm)->AttachCurrentThread(jni->jvm, __ENVP_CAST & jniEnv, NULL);
+    (*jni->jvm)->AttachCurrentThread(jni->jvm, &jniEnv, NULL);
 #if !defined(_WIN32)
     pthread_setspecific(tlsKey, &jniEnv);
 #endif