+++ pkgs/jnigen/lib/src/tools/android_sdk_tools.dart
@@ -4,6 +4,7 @@
 
 import 'dart:io';
 
+import 'package:jni_util/jni_util.dart' as jni_util;
 import 'package:path/path.dart';
 import 'package:yaml/yaml.dart';
 
@@ -482,7 +483,9 @@ tasks.register<DefaultTask>("$_gradleGetSourcesTaskName") {
     ProcessResult procRes;
     try {
       procRes = Process.runSync(gradleCommand, ['-q', taskPath],
-          workingDirectory: android, runInShell: true);
+          workingDirectory: android,
+          runInShell: true,
+          environment: jni_util.javaEnvironment);
     } finally {
       log.info('Restoring build scripts');
       origBuild.writeAsStringSync(
@@ -515,7 +518,7 @@ Execution failed for task ':gradle:compileGroovy'.
 > BUG! exception in phase 'semantic analysis' ...  Unsupported class file major version
 
 Then the JDK versions used by JNIgen and Flutter are not compatible. Try
-changing the default JDK version e.g. with `export JAVA_VERSION=11` on macOS and
+changing the default JDK version e.g. with `export JAVA_VERSION=17` on macOS and
 `sudo update-alternatives --config java` on Ubuntu.
 
 GRADLE OUTPUT:
@@ -561,6 +564,7 @@ ${procRes.stderr}
         ['build', 'apk', '--config-only'],
         workingDirectory: androidProject,
         runInShell: true,
+        environment: jni_util.javaEnvironment,
       );
       if (result.exitCode == 0) {
         success = true;