+++ pkgs/jnigen/test/large_java_test/large_java_test.dart
@@ -2,6 +2,9 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+@Tags(['bindings'])
+library;
+
 import 'dart:io';
 
 import 'package:jnigen/jnigen.dart';
@@ -31,12 +34,7 @@ Future<void> main() async {
             .where((f) => f.path.endsWith('.java'))
             .map((f) => p.relative(f.path, from: pkgDir))
             .toList();
-    final javacResult =
-        await Process.run('javac', [...javaFiles], workingDirectory: pkgDir);
-    expect(javacResult.exitCode, 0,
-        reason: 'Java compilation failed:\n'
-            'STDOUT: ${javacResult.stdout}\n'
-            'STDERR: ${javacResult.stderr}');
+    await runCommand('javac', [...javaFiles], workingDirectory: pkgDir);
 
     // Run JNIgen.
     final thisDir = Uri.directory(p.join(pkgDir, 'test', 'large_java_test'));