+++ pkgs/jnigen/test/simple_package_test/generate.dart
@@ -4,6 +4,7 @@
 
 import 'dart:io';
 
+import 'package:jni_util/jni_util.dart' as jni_util;
 import 'package:jnigen/jnigen.dart';
 import 'package:jnigen/src/logging/logging.dart';
 import 'package:logging/logging.dart';
@@ -71,7 +72,9 @@ final javaFiles = [
 ];
 
 void compileJavaSources(String workingDir, List<String> files) async {
-  final procRes = Process.runSync('javac', files, workingDirectory: workingDir);
+  final javac = jni_util.resolveJavaExecutable('javac');
+  final procRes = Process.runSync(javac, files,
+      workingDirectory: workingDir, environment: jni_util.javaEnvironment);
   if (procRes.exitCode != 0) {
     log.fatal('javac exited with ${procRes.exitCode}\n'
         '${procRes.stderr}');